chore(deps): Bump to Elysia 1.0.10
Elysia 1.0 introduces a breaking change for life cycle events.

> Starting from Elysia 1.0 RC, all lifecycle events including derive,
> and resolve will accept additional parameters
> { as: 'global' | 'local' } to specify if hook should be local or
> global which default to local

https://github.com/elysiajs/elysia/issues/513

elysia-htmx 1.0.9 does not specify such an option, and so any use of the
plugin results in the `hx` property not being typed as available in the
parent Elysia instance's handlers.

This commit updates dependencies to Elysia 1.0.10 and adds the as:
"global" option to allow the hx parameter to be available in parent
Elysia instances.

This mirrors how the official Bearer token plugin does it:
https://github.com/elysiajs/elysia-bearer/blob/main/src/index.ts#L56

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>